Fix file descriptor handling in bcron-exec.
There are two bugs fixed by this patch.

First, and most important, if a job had no MAILTO set, the file
descriptor for devnull would get closed for future jobs.

Second, if the input and output file descriptors to exec_cmd are the
same, it would get closed before being dup'ed to stdout.  This shouldn't
normally happen, but the reordering is still more correct.